+++ .gitignore
@@ -52,6 +52,7 @@ test/e2e/config.local.json
 
 # E2E test results and reports
 test-results/
+test/e2e/outputs/
 
 # Log files
 logs/
