+++ src/handlers/tool-configs/universal/field-mapper/transformers/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Field transformation modules - centralized exports
+ * Extracted from field-mapper.ts during Issue #529 modular refactoring
+ */
+
+// Core field mapping
+export { mapFieldName } from './field-mapper.js';
+
+// Value transformation
+export { transformFieldValue } from './value-transformer.js';
+
+// Record transformation
+export { mapRecordFields } from './record-transformer.js';
+
+// Task-specific transformation
+export { mapTaskFields } from './task-transformer.js';
